+++ AlphaFS.UnitTest/Directory Class/CreateDirectory.cs
@@ -193,9 +193,11 @@ namespace AlphaFS.UnitTest
             }
             catch (Exception ex)
             {
-               gotException = ex.GetType().Name.Equals("AlreadyExistsException", StringComparison.OrdinalIgnoreCase);
+               var exName = ex.GetType().Name;
+               gotException = exName.Equals("AlreadyExistsException", StringComparison.OrdinalIgnoreCase);
+               Console.WriteLine("\tCaught Exception: [{0}] Message: [{1}]", exName, ex.Message);
             }
-            Assert.IsTrue(gotException, "The exception was not caught, but is expected to.");
+            Assert.IsTrue(gotException, "The exception is not caught, but is expected to.");
          }
 
          Console.WriteLine();
@@ -221,9 +223,11 @@ namespace AlphaFS.UnitTest
          }
          catch (Exception ex)
          {
-            gotException = ex.GetType().Name.Equals("ArgumentException", StringComparison.OrdinalIgnoreCase);
+            var exName = ex.GetType().Name;
+            gotException = exName.Equals("ArgumentException", StringComparison.OrdinalIgnoreCase);
+            Console.WriteLine("\tCaught Exception: [{0}] Message: [{1}]", exName, ex.Message);
          }
-         Assert.IsTrue(gotException, "The exception was not caught, but is expected to.");
+         Assert.IsTrue(gotException, "The exception is not caught, but is expected to.");
 
          Console.WriteLine();
       }
@@ -244,9 +248,11 @@ namespace AlphaFS.UnitTest
          }
          catch (Exception ex)
          {
-            gotException = ex.GetType().Name.Equals("ArgumentException", StringComparison.OrdinalIgnoreCase);
+            var exName = ex.GetType().Name;
+            gotException = exName.Equals("ArgumentException", StringComparison.OrdinalIgnoreCase);
+            Console.WriteLine("\tCaught Exception: [{0}] Message: [{1}]", exName, ex.Message);
          }
-         Assert.IsTrue(gotException, "The exception was not caught, but is expected to.");
+         Assert.IsTrue(gotException, "The exception is not caught, but is expected to.");
 
          Console.WriteLine();
       }
@@ -270,9 +276,11 @@ namespace AlphaFS.UnitTest
          }
          catch (Exception ex)
          {
-            gotException = ex.GetType().Name.Equals(isNetwork ? "IOException" : "DirectoryNotFoundException", StringComparison.OrdinalIgnoreCase);
+            var exName = ex.GetType().Name;
+            gotException = exName.Equals(isNetwork ? "IOException" : "DirectoryNotFoundException", StringComparison.OrdinalIgnoreCase);
+            Console.WriteLine("\tCaught Exception: [{0}] Message: [{1}]", exName, ex.Message);
          }
-         Assert.IsTrue(gotException, "The exception was not caught, but is expected to.");
+         Assert.IsTrue(gotException, "The exception is not caught, but is expected to.");
 
          Console.WriteLine();
       }
@@ -295,9 +303,11 @@ namespace AlphaFS.UnitTest
          }
          catch (Exception ex)
          {
-            gotException = ex.GetType().Name.Equals("NotSupportedException", StringComparison.OrdinalIgnoreCase);
+            var exName = ex.GetType().Name;
+            gotException = exName.Equals("NotSupportedException", StringComparison.OrdinalIgnoreCase);
+            Console.WriteLine("\tCaught Exception: [{0}] Message: [{1}]", exName, ex.Message);
          }
-         Assert.IsTrue(gotException, "The exception was not caught, but is expected to.");
+         Assert.IsTrue(gotException, "The exception is not caught, but is expected to.");
 
          Console.WriteLine();
       }